requirements.txt
starlette
pydantic
rich>=13
pytest>=8
pytest-cov>=4.0.0
pytest-mock>=3.10.0
numpy>=1
setuptools>=68
bittensor>=9.10.1
bittensor-cli>=9.10.1
httpx>=0.25.0
python-dotenv>=1.0.0
